requirements.txt
telebot==0.0.5
python-dotenv==1.0.0
loguru~=0.7.0
SQLAlchemy~=2.0.20
pandas~=2.1.0
openpyxl==3.1.2
PyMySQL~=1.1.0